more typescript fixes
