+++ components/GcNav/DownloadOptions.vue
@@ -22,7 +22,7 @@
 </template>
 
 <script lang="ts" setup>
-import type { DownloadOptions } from "~/types/generic.types";
+import type { DownloadOptions } from "@/types/generic.types";
 
 import { XfIcon } from "xf-cmpt-lib";
