+++ nuxt.config.ts
@@ -53,7 +53,7 @@ export default defineNuxtConfig({
   },
   components: [
     {
-      path: "~/components",
+      path: "@/components",
       pathPrefix: false,
     },
   ],
